Landing Gear Stub completed. (Gear position and WOW) GW, Alt, Baro and IAS have an IO.

+++ Aero390Spoilers/Hydraulics/MainSystems.cs
@@ -17,10 +17,10 @@ namespace Hydraulics
         {
             AvailPressure = 3000;
         }
-        public float GetAvailPress()
+        public double GetAvailPress()
         {
             return AvailPressure;
         }
-        float AvailPressure; //psi
+        double AvailPressure; //psi
     }
 }
